:recycle: clean up code and rename classes for better context
