+++ libs/functions/bot-engine/main/src/lib/services/process-operation-block/block-type/jump-story-block.service.ts
@@ -8,8 +8,7 @@ import { StoryBlock } from "@app/model/convs-mgr/stories/blocks/main";
 
 import { BlockDataService } from "../../data-services/blocks.service";
 import { ConnectionsDataService } from "../../data-services/connections.service";
-import { EndUserDataService } from "../../data-services/end-user.service";
-import { IProcessNextBlock } from "../models/process-next-block.interface";
+import { IProcessOperationBlock } from "../models/process-operation-block.interface";
 
 
 /**
@@ -23,7 +22,7 @@ import { IProcessNextBlock } from "../models/process-next-block.interface";
  *  - Add function to get the first block in the new story if the block is not provided
  *  - Add function to jump to the same story, if the story is not provided
  */
-export class JumpStoryBlockService implements IProcessNextBlock
+export class JumpStoryBlockService implements IProcessOperationBlock
 {
   sideOperations: Promise<any>[] = [];
   userInput: string;
