+++ libs/functions/bot-engine/main/src/lib/services/process-operation-block/models/process-operation-block.interface.ts
@@ -1,7 +1,7 @@
 import { Cursor } from "@app/model/convs-mgr/conversations/admin/system";
 import { StoryBlock } from "@app/model/convs-mgr/stories/blocks/main";
 
-export interface IProcessNextBlock {
+export interface IProcessOperationBlock {
   sideOperations: Promise<any>[];
   handleBlock(storyBlock: StoryBlock, updatedCursor: Cursor, orgId: string, endUserId: string)
 }
