+++ libs/functions/bot-engine/main/src/lib/services/process-operation-block/process-operation-block.factory.ts
@@ -1,6 +1,5 @@
 import { HandlerTools } from "@iote/cqrs";
 
-import { Cursor } from "@app/model/convs-mgr/conversations/admin/system";
 import { StoryBlockTypes } from "@app/model/convs-mgr/stories/blocks/main";
 
 import { BlockDataService } from "../data-services/blocks.service";
@@ -8,14 +7,14 @@ import { ConnectionsDataService } from "../data-services/connections.service";
 import { WebhookBlockService } from "./block-type/webhook-block.service";
 
 import { JumpStoryBlockService } from "./block-type/jump-story-block.service";
-import { IProcessNextBlock } from "./models/process-next-block.interface";
+import { IProcessOperationBlock } from "./models/process-operation-block.interface";
 
 
-export class ProcessNextBlockFactory
+export class OperationBlockFactory
 {
   constructor(private _blockDataService: BlockDataService, private _connDataService: ConnectionsDataService, private tools: HandlerTools){}
 
-  resolve(storyBlockType: StoryBlockTypes): IProcessNextBlock
+  resolve(storyBlockType: StoryBlockTypes): IProcessOperationBlock
   {
     switch (storyBlockType) {
       case StoryBlockTypes.JumpBlock:
